resolvers unit tests (renamed queries)
